create model, & some request on db
