+++ .env-example
@@ -1,0 +1,4 @@
+DB_HOST=
+DB_USER=
+DB_PASSWORD=
+DB_NAME=
