+++ app.js
@@ -1,8 +1,24 @@
 const express = require("express");
 const logger = require("morgan");
 const cors = require("cors");
+const mongoose = require("mongoose");
+require("dotenv").config();
+
+const { DB_HOST } = process.env;
+console.log(DB_HOST);
 
 const contactsRouter = require("./routes/api/contacts");
+
+// const DB_HOST =
+//   "mongodb+srv://DenisMalniev:JgL0FmmYwZtwluJ3@cluster0.beiyw0s.mongodb.net/db-contacts?retryWrites=true&w=majority";
+
+mongoose
+  .connect(DB_HOST)
+  .then(() => console.log("Database connect successful"))
+  .catch((error) => {
+    console.log(error.message);
+    process.exit(1);
+  });
 
 const app = express();
