+++ middlewares/controllerWrapper.js
@@ -1,0 +1,10 @@
+const controllerWrapper = (controller) => {
+  return async (req, res, next) => {
+    try {
+      await controller(req, res, next);
+    } catch (error) {
+      next(error);
+    }
+  };
+};
+module.exports = controllerWrapper;
