+++ package.json
@@ -11,9 +11,11 @@
   "dependencies": {
     "cors": "2.8.5",
     "cross-env": "7.0.3",
+    "dotenv": "^16.0.1",
     "express": "4.17.1",
     "http-errors": "^2.0.0",
     "joi": "^17.6.0",
+    "mongoose": "^6.3.8",
     "morgan": "1.10.0",
     "uuid": "^8.3.2"
   },
