+++ readme.md
@@ -29,3 +29,8 @@
 - `npm run start:dev` &mdash; старт сервера в режимі розробки (development)
 - `npm run lint` &mdash; запустити виконання перевірки коду з eslint, необхідно виконувати перед кожним PR та виправляти всі помилки лінтера
 - `npm lint:fix` &mdash; та ж перевірка лінтера, але з автоматичними виправленнями простих помилок
+  //kissed03
+  //q$38a.g54Lge.4Y
+
+//DenisMalniev
+//JgL0FmmYwZtwluJ3
